+++ metacells/tools/group.py
@@ -3,8 +3,11 @@ Group
 -----
 """
 
+from hashlib import shake_128
+from textwrap import wrap
 from typing import Any
 from typing import Callable
+from typing import List
 from typing import Optional
 from typing import Union
 
@@ -54,6 +57,9 @@ def group_obs_data(
 
     The new data will contain only:
 
+    * A single observation for each group. The name of each observation will be 128 bit of the ``shake_128`` hash of the
+      names of the grouped members, in upper case hexadecimal (32 characters).
+
     * An ``X`` member holding the summed-per-group data.
 
     * A new ``grouped`` per-observation data which counts, for each group, the number
@@ -72,6 +78,7 @@ def group_obs_data(
 
     gdata = AnnData(summed_data)
     gdata.var_names = adata.var_names
+    gdata.obs_names = _obs_names(ut.to_numpy_vector(adata.obs_names), group_of_cells)
 
     ut.set_name(gdata, ut.get_name(adata))
     ut.set_name(gdata, name)
@@ -79,6 +86,19 @@ def group_obs_data(
     ut.set_o_data(gdata, "grouped", cell_counts, formatter=ut.sizes_description)
 
     return gdata
+
+
+def _obs_names(name_of_members: ut.NumpyVector, group_of_members: ut.NumpyVector) -> List[str]:
+    groups_count = np.max(group_of_members) + 1
+    name_of_groups: List[str] = []
+    for group_index in range(groups_count):
+        groups_mask = group_of_members == group_index
+        assert np.any(groups_mask)
+        hasher = shake_128()
+        for member_name in name_of_members[groups_mask]:
+            hasher.update(member_name.encode("utf8"))
+        name_of_groups.append(".".join(wrap(hasher.hexdigest(16).upper(), 4)))  # pylint: disable=too-many-function-args
+    return name_of_groups
 
 
 @ut.logged()