+++ HISTORY.rst
@@ -34,3 +34,9 @@ History
 * Fix edge case in computing noisy lonely genes.
 * Additional outliers certificates.
 * Stricter deviants detection policy
+
+0.9.0
+-----
+
+* Improved and published projection algorithm.
+* Add bystander genes (forbidden, but also ignored when computing deviant outlier cells).